requirements.txt
flask==3.0.3
pandas==2.2.3
gunicorn== 23.0.0
matplotlib==3.9.2
langchain-openai==0.2.3
langgraph==0.2.39
python-dotenv==1.0.1